Reduce CPU request to 2
Even though the project got quota for 2500 milli-CPUs per container/pod,
Kubernetes refuses to schedule the job if we want more than 2 CPUs.
Not sure if this is a Kubernetes problem (unlikely), or if the quota
change did not propagate to all nodes (equally unlikely), or something
else. At least, two CPUs are more than one.