Make LineMerger work on an interface
Our own PageSignalsScanner is essentially the same as a bufio.Scanner,
but without serialization. We could either write (quite a bit of) code
to serialize the lines to a buffer and then consume it right away in
a bufio.Scanner, or we abstract the functionality into a common interface.
Opting for the latter, because it's conceptually simpler. Probably also
more efficient (no serialization), but that's not the reason for the choice.