Make LineMerger read its input files in deterministic order
Before this change, it was not clearly defined which input stream
is put at the top of the min-heap if multiple streams happen to
contain the same line. This would not be a problem for correctness,
but the non-deterministic processing ordering makes debugging difficult
in presence of I/O problems.

https://github.com/brawer/wikidata-qrank/issues/40