ItemSignalBuilder: Download pageview files to local disk before processing
Apparently, the object storage system in Wikimedia's datacenter does not
like it when reading a file is taking more than about 25 minutes; at least
this might be an explanation for the non-deterministic I/O failures we
see then building item signals. Let's to fetch the pageview files to
a temporary directory (hopefully on local disk, but who knows) before
building item signals; it will be interesting to see whether this makes
any difference.

https://github.com/brawer/wikidata-qrank/issues/40